requirements.txt
requests
python-dotenv
python-twitter
python-environ
colored
pytz
tabulate
python-dateutil
beautifulsoup4
pytest
pandas